requirements.txt
torch>=2.5.1
torchvision>=0.20.1
pillow>=10.4.0
numpy>=2.2.6
typing_extensions>=4.15.0
matplotlib>=3.10.7